requirements.txt
fastapi
uvicorn
pydantic
httpx
azure-ai-openai
python-dotenv
spacy
en-core-web-sm